+++ src/cppref.ts
@@ -1,3 +1,3 @@
-export * from './classInfo';
 export * from './implementation';
 export * from './filePair';
+export * from './changeDeclaration';